rename _utils to stats
