+++ tests/test_stats.py
@@ -1,13 +1,13 @@
 import numpy as np
 
-from evident import _utils as utils
+from evident import stats
 
 
 def test_calc_pooled_std_two_groups():
     a = [1, 2, 3, 4, 5, 6]
     b = [2, 5, 3, 6, 8, 9]
 
-    calc_pooled_std = utils.calculate_pooled_stdev(a, b)
+    calc_pooled_std = stats.calculate_pooled_stdev(a, b)
     exp_pooled_std = 2.345208
     np.testing.assert_almost_equal(exp_pooled_std, calc_pooled_std, decimal=6)
 
@@ -17,7 +17,7 @@ def test_calc_pooled_std_three_groups():
     b = [2, 5, 3, 6, 8, 9]
     c = [0, 2, 2, 5, 1, 4]
 
-    calc_pooled_std = utils.calculate_pooled_stdev(a, b, c)
+    calc_pooled_std = stats.calculate_pooled_stdev(a, b, c)
     exp_pooled_std = 2.195955
     np.testing.assert_almost_equal(exp_pooled_std, calc_pooled_std, decimal=6)
 
@@ -26,7 +26,7 @@ def test_calc_cohens_d():
     a = [1, 2, 3, 4, 5, 6]
     b = [2, 5, 3, 6, 8, 9]
 
-    calc_cohen_d = utils.calculate_cohens_d(a, b)
+    calc_cohen_d = stats.calculate_cohens_d(a, b)
     exp_cohen_d = 0.852803
     np.testing.assert_almost_equal(exp_cohen_d, calc_cohen_d, decimal=6)
 
@@ -37,5 +37,5 @@ def test_calc_cohens_f_two_groups():
     b = [2, 5, 3, 6, 8, 9]
 
     exp_cohen_f = 0.852803 / 2
-    calc_cohen_f = utils.calculate_cohens_f(a, b)
+    calc_cohen_f = stats.calculate_cohens_f(a, b)
     np.testing.assert_almost_equal(exp_cohen_f, calc_cohen_f, decimal=6)
